Renaming location/code_location.py to location/coded_location.py and associated refactoring
