+++ mkdocs.yml
@@ -16,7 +16,7 @@ nav:
     - grids (package):
       - region_grid (module): api/grids/region_grid.md
     - location (package):
-      - code_location (module): api/location/code_location.md
+      - coded_location (module): api/location/coded_location.md
       - location (module): api/location/location.md
       - types (module): api/location/types.md
   - Contributing: contributing.md
